+++ src/main/java/online/kingdomkeys/kingdomkeys/magic/MagicReflect.java
@@ -5,7 +5,6 @@ import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
-import online.kingdomkeys.kingdomkeys.data.ModData;
 import online.kingdomkeys.kingdomkeys.client.sound.ModSounds;
 import online.kingdomkeys.kingdomkeys.data.PlayerData;
 import online.kingdomkeys.kingdomkeys.network.PacketHandler;
@@ -19,7 +18,7 @@ public class MagicReflect extends Magic {
 	@Override
 	public void magicUse(Player player, Player caster, int level, float fullMPBlastMult, LivingEntity lockOnEntity) {
 		PlayerData playerData = PlayerData.get(player);
-		playerData.setReflectTicks((int) (40 + (level * 5)), level);
+		playerData.setReflectTicks(40 + (level * 5), level);
 		PacketHandler.syncToAllAround(player, playerData);
 		player.swing(InteractionHand.MAIN_HAND);
 	}
